Create info
This is a basic example using the pygame library. It initializes the pygame mixer, loads a music file, plays it for 10 seconds, and then stops. Make sure to replace 'your_music_file.mp3' with the actual path to your music file.

Keep in mind that this is a simple example, and depending on your requirements, you may want to explore more advanced music programming libraries or frameworks.